fix: 수익률 계산 소수점 오차 개선
double 형에다 수익을 계산한 것을 더하면, 돈은 정수값인데, 실수형이 된다. 그러므로 long 형에다가 수익을 계산한 것을 더한다.

+++ src/main/java/lotto/model/util/CalculatorProfit.java
@@ -7,7 +7,7 @@ public class CalculatorProfit {
 
     public static double calProfitRate(Map<String, Integer> status, int buyAmount) {
 
-        double profitRate = 0.0;
+        long profitRate = 0L;
 
         for (String key : status.keySet()) {
             int correctCount = status.get(key);
@@ -20,6 +20,6 @@ public class CalculatorProfit {
         if (profitRate <= 0) {
             return 0;
         }
-        return profitRate / buyAmount * 100;
+        return ((double) profitRate / buyAmount) * 100;
     }
 }
